refactor(email_service/list_runs): move decode_key and encode_key to last_evaluated_key_util.py (SCRUM-120)
