+++ src/email_service/list_runs/lambda_function.py
@@ -9,6 +9,7 @@ from current_user_util import CurrentUserUtil
 from email_service_pagination_state_repository import (
     EmailServicePaginationStateRepository,
 )
+from last_evaluated_key_util import decode_key, encode_key
 from run_repository import RunRepository
 from time_util import get_current_utc_time, get_previous_year
 
@@ -24,26 +25,6 @@ class DecimalEncoder(json.JSONEncoder):
         if isinstance(o, Decimal):
             return float(o)
         return super(DecimalEncoder, self).default(o)
-
-
-def decode_key(encoded_key: str) -> dict[str, str]:
-    """Decode a base64 encoded key into a dictionary."""
-    try:
-        decoded_key = json.loads(base64.b64decode(encoded_key).decode("utf-8"))
-        return decoded_key
-    except (json.JSONDecodeError, base64.binascii.Error) as e:
-        logger.error("Invalid key format: %s", e)
-        raise ValueError(
-            "Invalid key format. It must be a valid base64 encoded JSON string."
-        ) from e
-
-
-def encode_key(decoded_key: dict[str, str]) -> str:
-    """Encode a dictionary into a base64 encoded string."""
-    encoded_key = base64.b64encode(json.dumps(decoded_key).encode("utf-8")).decode(
-        "utf-8"
-    )
-    return encoded_key
 
 
 def extract_query_params(event: dict[str, any]) -> dict[str, any]:
